SNOW-1229108 Add warning about buffer config when single buffer enabled (#896)
