+++ src/main/java/com/snowflake/kafka/connector/ConnectorConfigValidator.java
@@ -1,0 +1,14 @@
+package com.snowflake.kafka.connector;
+
+import java.util.Map;
+
+public interface ConnectorConfigValidator {
+
+  /**
+   * Validate input configuration
+   *
+   * @param config configuration Map
+   * @return connector name
+   */
+  String validateConfig(Map<String, String> config);
+}
